+++ EventViewer/Interfaces/IEventsDataService.cs
@@ -7,9 +7,11 @@ using System.Threading.Tasks;
 
 namespace EventViewer.Interfaces
 {
+
     public interface IEventsDataService
     {
         Task<List<EventData>> GetEventsData(DateTime? from = null, DateTime? to = null, Device device = null, string userId = null);
+        Task<List<CommandData>> GetCommandsData(DateTime? from = null, DateTime? to = null, Device device = null, string userId = null);
         Task<bool> CheckCredentials(string environment, string userId);
     }
 }