+++ EventViewer/Models/ApiLogin/ApiRefreshResponse.cs
@@ -6,8 +6,8 @@ using System.Threading.Tasks;
 
 namespace EventViewer.Models.ApiLogin
 {
-    public class ApiRefreshResponse
+    public class ApiRefreshResponse<T>
     {
-        public List<EventData> Data { get; set; } = new List<EventData>();
+        public List<T> Data { get; set; } = new List<T>();
     }
 }